[refactor] #59 유저 도메인 패키지 구성
