+++ src/main/java/_ganzi/codoc/auth/config/SecurityConfig.java
@@ -2,7 +2,7 @@ package _ganzi.codoc.auth.config;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
-import _ganzi.codoc.user.domain.UserStatus;
+import _ganzi.codoc.user.enums.UserStatus;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
