+++ src/main/java/_ganzi/codoc/user/enums/QuestStatus.java
@@ -1,4 +1,4 @@
-package _ganzi.codoc.user.domain;
+package _ganzi.codoc.user.enums;
 
 public enum QuestStatus {
     IN_PROGRESS,
