+++ src/main/java/_ganzi/codoc/user/enums/UserStatus.java
@@ -1,4 +1,4 @@
-package _ganzi.codoc.user.domain;
+package _ganzi.codoc.user.enums;
 
 public enum UserStatus {
     ACTIVE,
