+++ src/main/java/_ganzi/codoc/user/repository/AvatarRepository.java
@@ -1,5 +1,6 @@
-package _ganzi.codoc.user.domain;
+package _ganzi.codoc.user.repository;
 
+import _ganzi.codoc.user.domain.Avatar;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
