+++ src/main/java/_ganzi/codoc/user/repository/UserStatsRepository.java
@@ -1,5 +1,6 @@
-package _ganzi.codoc.user.domain;
+package _ganzi.codoc.user.repository;
 
+import _ganzi.codoc.user.domain.UserStats;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserStatsRepository extends JpaRepository<UserStats, Long> {}
