+++ src/main/java/_ganzi/codoc/user/service/UserService.java
@@ -2,9 +2,12 @@ package _ganzi.codoc.user.service;
 
 import _ganzi.codoc.user.api.dto.UserInitSurveyRequest;
 import _ganzi.codoc.user.domain.*;
+import _ganzi.codoc.user.enums.DailyGoal;
 import _ganzi.codoc.user.exception.AvatarNotFoundException;
 import _ganzi.codoc.user.exception.DuplicateNicknameException;
 import _ganzi.codoc.user.exception.UserNotFoundException;
+import _ganzi.codoc.user.repository.AvatarRepository;
+import _ganzi.codoc.user.repository.UserRepository;
 import _ganzi.codoc.user.service.dto.UserProfileResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
